fix: Allow lean4-mode in anonymous buffers
When putting Lean4 snippets in org mode buffers, on export it
calls (with-temp-buffer) to syntax highlight the code.
This fails when lean4-mode tries to find the associated lakefile.lean,
because there's no associated filesystem.

If the (buffer-find-file) function returns nil, we instead get the
filename of the buffer in the currently selected window.
This should be the original org-mode file, which is what we need.

This isn't a perfect solution, but it's a vast improvement.